phase4_adding_agents/services/etl-job/requirements.txt
boto3==1.35.0
psycopg2-binary==2.9.9
requests==2.32.3
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-requests==0.46b0
setuptools>=68,<70
wheel

